requirements/dev.txt
-r common.txt
pycodestyle==2.8.0
pylint==2.12.2
autopep8==1.6.0
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2